Conflicts:
	nodev.js

+++ nodev.js
@@ -179,6 +179,7 @@ watchFileChecker.check = function(cb) {
     return;
   }  
   fs.watch(watchFileName, function(event, filename) {
+    if (watchFileChecker.changeDetected) { return; }
     watchFileChecker.changeDetected = true;
     cb(true);
   });
@@ -423,7 +424,13 @@ function killNode() {
       // Force kill (/F) the whole child tree (/T) by PID (/PID 123)
       exec('taskkill /pid '+child.pid+' /T /F');
     } else {
-      child.kill('SIGUSR2');
+      killedAfterChange = true;
+  	  exec('pkill -P ' + child.pid, function (error, stdout, stderr) {
+  	    if (error !== null) {
+          killedAfterChange = false;
+  		    child.kill('SIGUSR2');
+  	    }
+  	  });
     }
   } else {
     startNode();
